python_mysql_migration/requirements.txt
pymongo
pandas
python-dotenv